requirements.txt
robotframework==7.3.2
robotframework-seleniumlibrary==6.7.1
selenium==4.34.2
Pillow==10.4.0